# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Mente Livre API'
  description: 'API REST para el sistema de gestión de terapeutas y citas de Mente Livre'
  contact:
    name: 'Mente Livre'
    email: mentelivre.pe@gmail.com
  version: 1.0.0
servers:
  -
    url: 'https://backend.mentelivre.org/'
    description: 'Servidor de Desarrollo Local'
paths:
  /appointments:
    get:
      tags:
        - Appointments
      summary: 'Listar citas'
      operationId: getAppointments
      parameters:
        -
          name: therapist_id
          in: query
          required: false
          schema:
            type: string
        -
          name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - pending
              - confirmed
              - completed
              - cancelled
              - pending_payment
              - payment_review
        -
          name: date_from
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: date_to
          in: query
          required: false
          schema:
            type: string
            format: date
        -
          name: patient_email
          in: query
          required: false
          schema:
            type: string
            format: email
      responses:
        '200':
          description: 'Lista de citas'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
    post:
      tags:
        - Appointments
      summary: 'Crear nueva cita'
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: 'Cita creada exitosamente'
          content:
            application/json:
              schema:
                properties:
                  id: { type: string }
                  message: { type: string }
                type: object
        '400':
          description: 'Datos inválidos'
  '/appointments/{id}':
    get:
      tags:
        - Appointments
      summary: 'Obtener cita por ID'
      operationId: getAppointment
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Detalle de la cita'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          description: 'Cita no encontrada'
    put:
      tags:
        - Appointments
      summary: 'Actualizar cita'
      operationId: updateAppointment
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: 'Cita actualizada exitosamente'
        '404':
          description: 'Cita no encontrada'
    delete:
      tags:
        - Appointments
      summary: 'Eliminar cita'
      operationId: deleteAppointment
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Cita eliminada exitosamente'
        '404':
          description: 'Cita no encontrada'
...
components:
  schemas:
    Appointment:
      properties:
        id:
          type: string
        therapist_id:
          type: string
        therapist_name:
          type: string
        patient_email:
          type: string
        patient_name:
          type: string
        patient_phone:
          type: string
          nullable: true
        consultation_reason:
          type: string
          nullable: true
        appointment_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        status:
          type: string
          enum:
            - pending
            - confirmed
            - completed
            - cancelled
            - pending_payment
            - payment_review
        pricing_tier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    AppointmentCreate:
      required:
        - therapist_id
        - patient_email
        - patient_name
        - appointment_date
        - start_time
        - end_time
      properties:
        therapist_id:
          type: string
        patient_email:
          type: string
          format: email
        patient_name:
          type: string
        patient_phone:
          type: string
          nullable: true
        consultation_reason:
          type: string
          nullable: true
        appointment_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        status:
          type: string
          default: pending
          enum:
            - pending
            - confirmed
        pricing_tier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        original_price:
          type: number
          nullable: true
        discount_applied:
          type: number
          nullable: true
        final_price:
          type: number
          nullable: true
      type: object
    AppointmentUpdate:
      properties:
        status:
          type: string
          enum:
            - pending
            - confirmed
            - completed
            - cancelled
            - pending_payment
            - payment_review
        appointment_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        consultation_reason:
          type: string
        notes:
          type: string
        patient_name:
          type: string
        patient_phone:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token JWT de autenticación. Incluir en el header: Authorization: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...